# Conflicts:
#	compilerproject.iml
#	src/antlr/DartParser.interp
#	src/antlr/DartParser.java
#	src/antlr/DartParserBaseVisitor.java
#	src/antlr/DartParserVisitor.java
#	src/grammars/DartParser.g4
#	src/interfaces/IAntlrObjectFactory.java
#	src/properties/Children.java
#	src/properties/HeightProperty.java
#	src/properties/WidthProperty.java
#	src/visitors/AntlrObjectFactory.java
#	src/visitors/AntlrToProperty.java
#	src/visitors/AntlrToWidget.java

+++ src/properties/Right.java
@@ -1,14 +1,14 @@
 package properties;
 
 public class Right extends Property{
-    private final int value;
+    private final double value;
 
-    public Right(int value) {
+    public Right(double value) {
         super("right");
         this.value = value;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 }
